Put transformed Gen constructors in Test.Util.QuickCheck
