+++ ouroboros-consensus-test/src/Test/Util/ChainUpdates.hs
@@ -15,6 +15,7 @@ module Test.Util.ChainUpdates (
 import           Control.Monad.State.Strict
 
 import           Test.QuickCheck
+import           Test.Util.QuickCheck (frequency')
 
 import           Ouroboros.Network.MockChain.Chain (Chain (Genesis))
 import qualified Ouroboros.Network.MockChain.Chain as Chain
@@ -163,17 +164,6 @@ genChainUpdateState updateBehavior securityParam n =
       genAddBlock Invalid
       genSwitchFork (pure 1)
 
--- | Variant of 'frequency' that allows for transformers of 'Gen'
-frequency' :: (MonadTrans t, Monad (t Gen)) => [(Int, t Gen a)] -> t Gen a
-frequency' [] = error "frequency' used with empty list"
-frequency' xs0 = lift (choose (1, tot)) >>= (`pick` xs0)
-  where
-    tot = sum (map fst xs0)
-
-    pick n ((k,x):xs)
-      | n <= k    = x
-      | otherwise = pick (n-k) xs
-    pick _ _  = error "pick used with empty list"
 
 -- | Test that applying the generated updates gives us the same chain
 -- as @cusCurrentChain@.
